refactor: splits the two parsers into separate files
