+++ src/main.ts
@@ -1,15 +1,13 @@
 import type { IChange, IOptions } from "../types/watskeburt.js";
-import {
-  convertDiffLines,
-  convertStatusLines,
-} from "./convert-to-change-object.js";
+import { parseDiffLines } from "./parse-diff-lines.js";
+import { parseStatusLines } from "./parse-status-lines.js";
 import * as primitives from "./git-primitives.js";
 import format from "./formatters/format.js";
 
 export async function list(
   pOldRevision?: string,
   pNewRevision?: string,
-  pOptions?: IOptions,
+  pOptions?: IOptions
 ): Promise<IChange[] | string> {
   const lOldRevision: string = pOldRevision || (await primitives.getSHA());
   const lOptions: IOptions = pOptions || {};
@@ -21,13 +19,13 @@ export async function list(
     !lOptions.trackedOnly ? primitives.getStatusShort() : "",
   ]);
 
-  let lChanges = convertDiffLines(lDiffLines);
+  let lChanges = parseDiffLines(lDiffLines);
 
   if (!lOptions.trackedOnly) {
     lChanges = lChanges.concat(
-      convertStatusLines(lStatusLines).filter(
-        ({ changeType }) => changeType === "untracked",
-      ),
+      parseStatusLines(lStatusLines).filter(
+        ({ changeType }) => changeType === "untracked"
+      )
     );
   }
   return format(lChanges, lOptions.outputType);
